Extract team member project view in different component
